Fix mergeRecursive when catalog items in a group have the same name but are different datasets (distinguished by id)
